move constants to typings as well
