+++ lib/actions/config.ts
@@ -1,4 +1,4 @@
-import {CONFIG_LOAD, CONFIG_RELOAD} from '../constants/config';
+import {CONFIG_LOAD, CONFIG_RELOAD} from '../../typings/constants/config';
 import type {HyperActions} from '../../typings/hyper';
 import type {configOptions} from '../../typings/config';
