+++ lib/actions/updater.ts
@@ -1,4 +1,4 @@
-import {UPDATE_INSTALL, UPDATE_AVAILABLE} from '../constants/updater';
+import {UPDATE_INSTALL, UPDATE_AVAILABLE} from '../../typings/constants/updater';
 import rpc from '../rpc';
 import type {HyperActions} from '../../typings/hyper';
