+++ lib/reducers/term-groups.ts
@@ -1,9 +1,9 @@
 import {v4 as uuidv4} from 'uuid';
 import type {Immutable as ImmutableType} from 'seamless-immutable';
 import Immutable from 'seamless-immutable';
-import {TERM_GROUP_EXIT, TERM_GROUP_RESIZE} from '../constants/term-groups';
-import type {SessionAddAction} from '../constants/sessions';
-import {SESSION_ADD, SESSION_SET_ACTIVE} from '../constants/sessions';
+import {TERM_GROUP_EXIT, TERM_GROUP_RESIZE} from '../../typings/constants/term-groups';
+import type {SessionAddAction} from '../../typings/constants/sessions';
+import {SESSION_ADD, SESSION_SET_ACTIVE} from '../../typings/constants/sessions';
 import findBySession from '../utils/term-groups';
 import {decorateTermGroupsReducer} from '../utils/plugins';
 import type {ITermGroup, ITermState, ITermGroups, ITermGroupReducer, Mutable} from '../../typings/hyper';
