+++ lib/reducers/ui.ts
@@ -1,7 +1,7 @@
 import type {Immutable as ImmutableType} from 'seamless-immutable';
 import Immutable from 'seamless-immutable';
 import {decorateUIReducer} from '../utils/plugins';
-import {CONFIG_LOAD, CONFIG_RELOAD} from '../constants/config';
+import {CONFIG_LOAD, CONFIG_RELOAD} from '../../typings/constants/config';
 import {
   UI_FONT_SIZE_SET,
   UI_FONT_SIZE_RESET,
@@ -11,8 +11,8 @@ import {
   UI_WINDOW_GEOMETRY_CHANGED,
   UI_ENTER_FULLSCREEN,
   UI_LEAVE_FULLSCREEN
-} from '../constants/ui';
-import {NOTIFICATION_MESSAGE, NOTIFICATION_DISMISS} from '../constants/notifications';
+} from '../../typings/constants/ui';
+import {NOTIFICATION_MESSAGE, NOTIFICATION_DISMISS} from '../../typings/constants/notifications';
 import {
   SESSION_ADD,
   SESSION_RESIZE,
@@ -20,8 +20,8 @@ import {
   SESSION_PTY_EXIT,
   SESSION_SET_ACTIVE,
   SESSION_SET_CWD
-} from '../constants/sessions';
-import {UPDATE_AVAILABLE} from '../constants/updater';
+} from '../../typings/constants/sessions';
+import {UPDATE_AVAILABLE} from '../../typings/constants/updater';
 import type {uiState, Mutable, IUiReducer} from '../../typings/hyper';
 import {release} from 'os';
