+++ tsconfig.json
@@ -8,7 +8,7 @@
     "./lib/",
     "./test/",
     "./cli/",
-    "./typings"
+    "./typings/"
   ],
   "references": [
     {
