+++ typings/constants/tabs.d.ts
